Merge stable Kodi Omega code

+++ xbmc/addons/kodi-dev-kit/include/kodi/addon-instance/Inputstream.h
@@ -152,10 +152,7 @@ public:
   std::string GetMimeType() const { return m_cStructure->m_mimeType; }
 
   /// @brief Amount of available properties.
-  unsigned int GetPropertiesAmount() const
-  {
-    return m_cStructure->m_nCountInfoValues;
-  }
+  unsigned int GetPropertiesAmount() const { return m_cStructure->m_nCountInfoValues; }
 
   /// @brief List of available properties-
   const std::map<std::string, std::string> GetProperties() const
@@ -652,7 +649,7 @@ public:
   {
     m_extraData = extraData;
     m_cStructure->m_ExtraData = m_extraData.data();
-    m_cStructure->m_ExtraSize = m_extraData.size();
+    m_cStructure->m_ExtraSize = static_cast<unsigned int>(m_extraData.size());
   }
 
   /// @brief Additional data where can needed on streams.
@@ -669,7 +666,7 @@ public:
     }
 
     m_cStructure->m_ExtraData = m_extraData.data();
-    m_cStructure->m_ExtraSize = m_extraData.size();
+    m_cStructure->m_ExtraSize = static_cast<unsigned int>(m_extraData.size());
   }
 
   /// @brief To get with @ref SetExtraData changed values.
@@ -700,7 +697,7 @@ public:
   {
     m_extraData.clear();
     m_cStructure->m_ExtraData = m_extraData.data();
-    m_cStructure->m_ExtraSize = m_extraData.size();
+    m_cStructure->m_ExtraSize = static_cast<unsigned int>(m_extraData.size());
   }
 
   /// @brief RFC 5646 language code (empty string if undefined).
@@ -1781,7 +1778,7 @@ private:
 
   void SetAddonStruct(KODI_ADDON_INSTANCE_STRUCT* instance)
   {
-    int api[3] = { 0, 0, 0 };
+    int api[3] = {0, 0, 0};
     sscanf(GetInstanceAPIVersion().c_str(), "%d.%d.%d", &api[0], &api[1], &api[2]);
 
     instance->hdl = this;
@@ -1853,7 +1850,6 @@ private:
     InputstreamCapabilities caps(capabilities);
     static_cast<CInstanceInputStream*>(instance->toAddon->addonInstance)->GetCapabilities(caps);
   }
-
 
   // IDemux
   inline static bool ADDON_GetStreamIds(const AddonInstance_InputStream* instance,
